# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "BooleanValue": {
        "additionalProperties": false,
        "properties": {
          "type": {
            "const": "BOOLEAN",
            "title": "Type",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "boolean"
          }
        },
        "required": [
          "type",
          "value"
        ],
        "title": "BooleanValue",
        "type": "object"
      },
...
      "Envelope_Page_Profile__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Page_Profile_"
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "Envelope[Page[Profile]]",
        "type": "object"
      },
...
      "Envelope_Profile_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Profile"
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "Envelope[Profile]",
        "type": "object"
      },
...
      "Fact": {
        "additionalProperties": false,
        "properties": {
          "attribute": {
            "$ref": "#/components/schemas/FactAttribute"
          },
          "confirmed_at": {
            "format": "date-time",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(\\.[0-9]{1,6})?Z$",
            "title": "Confirmed At",
            "type": "string"
          },
          "conflict": {
            "title": "Conflict",
            "type": "boolean"
          },
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array",
            "uniqueItems": true
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "value": {
            "discriminator": {
              "mapping": {
                "BOOLEAN": "#/components/schemas/BooleanValue",
                "COUNTRY_SET": "#/components/schemas/CountrySetValue",
                "DATE": "#/components/schemas/DateValue",
                "EXPERIENCE": "#/components/schemas/ExperienceValue",
                "GPA": "#/components/schemas/GpaValue",
                "LANGUAGE_TESTS": "#/components/schemas/LanguageTestsValue",
                "STRING": "#/components/schemas/StringValue",
                "STRING_SET": "#/components/schemas/StringSetValue",
                "UNKNOWN": "#/components/schemas/UnknownValue"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/StringValue"
              },
              {
                "$ref": "#/components/schemas/BooleanValue"
              },
              {
                "$ref": "#/components/schemas/CountrySetValue"
              },
              {
                "$ref": "#/components/schemas/StringSetValue"
              },
              {
                "$ref": "#/components/schemas/GpaValue"
              },
              {
                "$ref": "#/components/schemas/DateValue"
              },
              {
                "$ref": "#/components/schemas/ExperienceValue"
              },
              {
                "$ref": "#/components/schemas/LanguageTestsValue"
              },
              {
                "$ref": "#/components/schemas/UnknownValue"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "attribute",
          "value",
          "evidence_ids",
          "id",
          "provenance",
          "confirmed_at",
          "valid_from",
          "valid_until",
          "conflict"
        ],
        "title": "Fact",
        "type": "object"
      },
      "FactAttribute": {
        "enum": [
          "location.country",
          "citizenship.countries",
          "work_authorization.countries",
          "education.enrolled",
          "education.level",
          "education.field",
          "education.institution",
          "education.graduation",
          "date_of_birth",
          "education.gpa",
          "skills",
          "experience",
          "language_tests"
        ],
        "title": "FactAttribute",
        "type": "string"
      },
...
      "FactInput": {
        "additionalProperties": false,
        "properties": {
          "attribute": {
            "$ref": "#/components/schemas/FactAttribute"
          },
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array",
            "uniqueItems": true
          },
          "value": {
            "discriminator": {
              "mapping": {
                "BOOLEAN": "#/components/schemas/BooleanValue",
                "COUNTRY_SET": "#/components/schemas/CountrySetValue",
                "DATE": "#/components/schemas/DateValue",
                "EXPERIENCE": "#/components/schemas/ExperienceValue",
                "GPA": "#/components/schemas/GpaValue",
                "LANGUAGE_TESTS": "#/components/schemas/LanguageTestsValue",
                "STRING": "#/components/schemas/StringValue",
                "STRING_SET": "#/components/schemas/StringSetValue",
                "UNKNOWN": "#/components/schemas/UnknownValue"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/StringValue"
              },
              {
                "$ref": "#/components/schemas/BooleanValue"
              },
              {
                "$ref": "#/components/schemas/CountrySetValue"
              },
              {
                "$ref": "#/components/schemas/StringSetValue"
              },
              {
                "$ref": "#/components/schemas/GpaValue"
              },
              {
                "$ref": "#/components/schemas/DateValue"
              },
              {
                "$ref": "#/components/schemas/ExperienceValue"
              },
              {
                "$ref": "#/components/schemas/LanguageTestsValue"
              },
              {
                "$ref": "#/components/schemas/UnknownValue"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "attribute",
          "value",
          "evidence_ids"
        ],
        "title": "FactInput",
        "type": "object"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "path": {
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Page_Profile_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Profile"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[Profile]",
        "type": "object"
      },
...
      "PatchProfile": {
        "additionalProperties": false,
        "properties": {
          "base_profile_version_id": {
            "format": "uuid",
            "title": "Base Profile Version Id",
            "type": "string"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/FactInput"
            },
            "maxItems": 100,
            "title": "Changes",
            "type": "array"
          },
          "remove_attributes": {
            "items": {
              "$ref": "#/components/schemas/FactAttribute"
            },
            "maxItems": 100,
            "title": "Remove Attributes",
            "type": "array"
          }
        },
        "required": [
          "base_profile_version_id",
          "changes",
          "remove_attributes"
        ],
        "title": "PatchProfile",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "pattern": "^urn:benefitbridge:problem:[a-z0-9-]+$",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "request_id",
          "errors",
          "retryable"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Profile": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "maxItems": 100,
            "title": "Facts",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(\\.[0-9]{1,6})?Z$",
            "title": "Updated At",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          },
          "version_number": {
            "minimum": 1,
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version_id",
          "version_number",
          "facts",
          "updated_at"
        ],
        "title": "Profile",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BenefitBridge",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/profile": {
      "get": {
        "operationId": "get_profile",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Profile_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile"
      },
      "patch": {
        "operationId": "patch_profile",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[!-~]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Profile_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Profile"
      }
    },
    "/api/v1/profile/versions": {
      "get": {
        "operationId": "list_profile_versions",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Page_Profile__"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Profile Versions"
      }
    },
    "/api/v1/profile/versions/{profile_version_id}": {
      "get": {
        "operationId": "get_profile_version",
        "parameters": [
          {
            "in": "path",
            "name": "profile_version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Profile_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile Version"
      }
    }
  }
}